Manager 로직 수정. Data Manager 추가, EnemyTB/EnemyData 프로퍼티

+++ Assets/Script/Interface/IDamagable.cs
@@ -4,6 +4,6 @@ public interface IDamagable
 {
     // 데미지를 입을 수 있는 오브젝트 들을 위한 메서드 및 멤버
 
-    public void TakeDamage(int _value);                    // 데미지를 입음.
+    public void TakeDamage(float _value);                    // 데미지를 입음.
 
 }
